split code up into component classes, burgerMenu improvements
